+++ src/Request/CancelRequest.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Nexy\PayboxDirect\Request;
+
+/**
+ * @author Sullivan Senechal <soullivaneuh@gmail.com>
+ */
+final class CancelRequest extends AbstractNumberedReferencedTransactionRequest
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function getRequestId()
+    {
+        return RequestInterface::CANCEL;
+    }
+}
