+++ tests/Request/CancelRequestTest.php
@@ -2,25 +2,25 @@
 
 namespace Nexy\PayboxDirect\Tests\Request;
 
-use Nexy\PayboxDirect\Request\AuthorizeRequest;
-use Nexy\PayboxDirect\Request\DebitRequest;
+use Nexy\PayboxDirect\Request\AuthorizeAndCaptureRequest;
+use Nexy\PayboxDirect\Request\CancelRequest;
 
 /**
  * @author Sullivan Senechal <soullivaneuh@gmail.com>
  */
-final class DebitRequestTest extends AbstractTransactionRequestTest
+final class CancelRequestTest extends AbstractTransactionRequestTest
 {
     public function testCall()
     {
-        $request = new AuthorizeRequest(
+        $request = new AuthorizeAndCaptureRequest(
             $this->generateReference(),
-            20000,
+            50000,
             $this->getCreditCardSerial(),
             $this->getCreditCardValidDate()
         );
         $response = $this->paybox->request($request);
 
-        $request = new DebitRequest($this->generateReference(), 20000, $response->getTransactionNumber(), $response->getCallNumber());
+        $request = new CancelRequest($this->generateReference(), 50000, $response->getTransactionNumber(), $response->getCallNumber());
         $response = $this->paybox->request($request);
 
         $this->assertSame(0, $response->getCode(), $response->getComment());
@@ -31,15 +31,15 @@ final class DebitRequestTest extends AbstractTransactionRequestTest
      */
     protected function createBaseRequest()
     {
-        $request = new AuthorizeRequest(
+        $request = new AuthorizeAndCaptureRequest(
             $this->generateReference(),
-            20042,
+            50042,
             $this->getCreditCardSerial(),
             $this->getCreditCardValidDate()
         );
         $response = $this->paybox->request($request);
 
-        $request = new DebitRequest($this->generateReference(), 20042, $response->getTransactionNumber(), $response->getCallNumber());
+        $request = new CancelRequest($this->generateReference(), 50042, $response->getTransactionNumber(), $response->getCallNumber());
 
         return $request;
     }
